refactor optimistic iouReport to ReportUtils
